+++ magda-web-client/src/config.js
@@ -102,7 +102,11 @@ export const config = {
             links: [
                 [
                     "API Docs",
-                    "http://search.data.gov.au/api/v0/registry/swagger/index.html"
+                    "https://search.data.gov.au/api/v0/registry/swagger/index.html"
+                ],
+                [
+                    "Powered by Magda",
+                    "https://github.com/TerriaJS/magda/"
                 ]
             ]
         },